--- conui.h
@@ -1,10 +1,0 @@
-//
-// Created by Phosphorus15 on 2019/1/19.
-//
-
-#ifndef TEACHINGMANAGEMENT_CONUI_H
-#define TEACHINGMANAGEMENT_CONUI_H
-
-
-
-#endif //TEACHINGMANAGEMENT_CONUI_H
