+++ uis/edit_course.h
@@ -9,7 +9,7 @@
 
 #include <iostream>
 #include <conio.h>
-#include "console.h"
+#include "../console.h"
 
 using namespace console::color;
