+++ uis/manage_course.h
@@ -8,7 +8,7 @@
 
 #include <iostream>
 #include <iomanip>
-#include "console.h"
+#include "../console.h"
 #include <cstdio>
 #include <conio.h>
 
@@ -16,7 +16,7 @@
 #pragma clang diagnostic push
 #pragma clang diagnostic ignored "-Wmissing-noreturn"
 extern "C" {
-#include "sqlite3/sqlite3.h"
+#include "../sqlite3/sqlite3.h"
 }
 
 using namespace console::color;
