+++ uis/teacher_append.h
@@ -1,20 +1,18 @@
 //
-// Created by Bario on 2019/1/21.
+// Created by Bario on 2019/1/22.
 //
 
-#ifndef TEACHINGMANAGEMENT_ADD_STUDENT_INFORMATION_H
-#define TEACHINGMANAGEMENT_ADD_STUDENT_INFORMATION_H
+#ifndef TEACHINGMANAGEMENT_ADD_TEACHER_INFORMATION_H
+#define TEACHINGMANAGEMENT_ADD_TEACHER_INFORMATION_H
 
-#endif //TEACHINGMANAGEMENT_ADD_STUDENT_INFORMATION_H
-
-
+#endif //TEACHINGMANAGEMENT_ADD_TEACHER_INFORMATION_H
 #include <iostream>
 #include <conio.h>
-#include "console.h"
+#include "../console.h"
 
 using namespace console::color;
 
-namespace add_student_information {
+namespace add_teacher_information {
 
     bool confirm(std::string msg, int line) {
         console::setcursor(0, line);
@@ -39,41 +37,32 @@ namespace add_student_information {
             }
         }
     }
-    struct Student{
+    struct Teacher{
         std::string id;
-        std::string clazz;
         std::string name;
         char gender;
-        std::string birthday;
-        std::string source;
-        long long tel;
+        std::string professional_title;
     };
-    void save_add_student(Student student) {
+    void save_add_Teacher(Teacher teacher) {
         // TODO
     }
-    void add_student_information(){
+    void add_teacher_information(){
         console::clearscreen();
-        Student student;
-    std::cout<<"---请输入学生信息：";
-    std::cout<<"学号：";
-    std::cin>>student.id;
-    std::cout<<"班号：";
-    std::cin>>student.clazz;
-    std::cout<<"姓名：";
-    std::cin>>student.name;
-    std::cout<<"性别:";
-    std::cin>>student.gender;
-    std::cout<<"出生日期:";
-    std::cin>>student.birthday;
-    std::cout<<"生源:";
-    std::cin>>student.source;
-    std::cout<<"电话:";
-    std::cin>>student.tel;
-        save_add_student(student);
+        Teacher teacher;
+        std::cout<<"---请输入老师信息：";
+        std::cout<<"工号：";
+        std::cin>>teacher.id;
+        std::cout<<"姓名：";
+        std::cin>>teacher.name;
+        std::cout<<"性别:";
+        std::cin>>teacher.gender;
+        std::cout<<"教师职称:";
+        std::cin>>teacher.professional_title;
+        save_add_Teacher(teacher);
         std::cout <<setc(yellow)<< "保存成功!" << std::endl;
-        save_add_student(student);
+        save_add_Teacher(teacher);
         if (confirm("保存成功，是否继续录入?", 10)) {
-            add_student_information();
+            add_teacher_information();
         } else {
             console::clearscreen();
         }
