+++ uis/teacher_delete.h
@@ -8,7 +8,7 @@
 #endif //TEACHINGMANAGEMENT_DELETE_TEACHER_INFORMATION_H
 #include <iostream>
 #include <conio.h>
-#include "console.h"
+#include "../console.h"
 
 using namespace console::color;
 
@@ -29,7 +29,7 @@ namespace  delete_teacher_information{
 
     void delete_teacher_information(){
         console::clearscreen();
-        std::cout << "请输入您要选择的教师姓名: ";
+        std::cout << "请输入您要删除的教师姓名: ";
         std::string name;
         std::cin >> name;
         Teacher teacher=get_teacher(name);
